Working JTAG connection
This patch add jtag_test function that allow to check by analizing
output if connection is correct (you could shift DR in the loop and see
IDCODEs and leater bits that where shifted in)

TDO is changed to input insted of output (it is our input, TDO is output
for debuged device).

Low level JTAG logic is changed. Now it holds TCK up until next jtag_tdi
request is issued. TCK is set to low in the same moment when TDI and TMS
are setup. TDI is read right after rising edge of TCK.