Tune JTAG clock cycle
Current implementation doesn't need delay, because code it self need
1 micro second per TCK cycle. This gives 1 MHz so it is correct for
targets with 6 MHz CPU clocks and faster.

This patch add special delay routine when TCKWAIT is 0. If TCKWAIT is
non zero it is interpreted as delay in microseconds. This way it is easy
to set longer TCK cycle if needed.